Frontend logic for the time and date selectors
