+++ apps/site/assets/js/app.js
@@ -29,7 +29,7 @@ import timetableStick from "./timetable-stick";
 import menuClose from "./menu-close";
 import datePicker from "./date-picker";
 import toggleBtn from "./toggle-on-click";
-import * as TripPlannerTimeControls from "./trip-planner-time-controls";
+import * as TimeControls from "./time-controls/time-controls";
 import * as TripPlannerLocationControls from "./trip-planner-location-controls";
 import stopBubbles from "./stop-bubbles";
 import search from "./search";
@@ -357,7 +357,7 @@ globalSearch.init();
 homepageSearch.init();
 embeddedSearchInit();
 TripPlannerLocationControls.init();
-TripPlannerTimeControls.init();
+TimeControls.init();
 fullstory();
 channels();
 CRTrains();
